+++ backend/src/api/medias/medias.module.ts
@@ -4,11 +4,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { FeedMediaEntity } from '@/models/entities/feed-media.entity';
 import { MediasRepository } from '@/models/repositories/medias.repository';
 
+import { MediasController } from './medias.controller';
 import { MediasService } from './medias.service';
 
 @Module({
 	imports: [TypeOrmModule.forFeature([FeedMediaEntity])],
-	controllers: [],
+	controllers: [MediasController],
 	providers: [MediasService, MediasRepository],
 	exports: [MediasService],
 })
